refactor: decouple package optimize
